processing for IntentSpec
