+++ FortranAst/src/pt/up/fe/specs/fortran/ast/nodes/type/attributes/IntentSpec.java
@@ -1,14 +1,14 @@
-package pt.up.fe.specs.fortran.ast.nodes.utils;
+package pt.up.fe.specs.fortran.ast.nodes.type.attributes;
 
 import org.suikasoft.jOptions.Datakey.DataKey;
 import org.suikasoft.jOptions.Datakey.KeyFactory;
 import org.suikasoft.jOptions.Interfaces.DataStore;
 import pt.up.fe.specs.fortran.ast.nodes.FortranNode;
-import pt.up.fe.specs.fortran.ast.nodes.utils.enums.IntentKind;
+import pt.up.fe.specs.fortran.ast.nodes.type.attributes.enums.IntentKind;
 
 import java.util.Collection;
 
-public class IntentSpec extends FortranNode {
+public class IntentSpec extends AttributeSpecifier {
 
     public final static DataKey<IntentKind> KIND = KeyFactory.enumeration("kind", IntentKind.class);
