+++ FortranParser/src/pt/up/fe/specs/fortran/parser/FlangName.java
@@ -125,6 +125,7 @@ public enum FlangName implements StringProvider {
     EXPLICIT_SHAPE_SPEC,
     ALLOCATABLE,
     ASYNCHRONOUS,
+    INTENT_SPEC,
 
 
     // OTHER
